Move all MoneyTransfersControllers in same folder and extract same methods in BaseMoneyTransferController.cs
